New layout and barcode layout for label printing
Added one normal new layout and a new barcode layout for ITDB. We used
them with Avery Zweckform sheets and they work really good.

Below some screenshots of how the labels will look and what changes at
the interface. Please ignore the borders around the labels ... had not
the time to make it fit into the label, so I just used the existing
labels without any modification which are obviously not suitable with
adjustment.